Fix issues with event call on release
Fix crash when there is no Biometric on the device
